# Use db.Get when look up metadata in pebble