<!-- Optional: Provide additional context (beyond the PR title). -->

<!-- Optional: link a GitHub issue.
     Example: "Fixes #123" will auto-close #123 when the PR is merged. -->

**Related issues**: N/A
This PR created a new method pebble.GetProto; which uses db.Get to get the bytes
and then unmarshal it to proto.

I used GetProto inside the lookupMetadataAndVersion method. It is faster and
uses less alloc per op. Also, we no longer need to worry about the order of
iterator creation and locking the key. 

Here is the benchmark test results:
```
                          │ /home/lulu/benchmark-pebble-master.log │ /home/lulu/benchmark-pebble-get-proto.log │
                          │                 sec/op                 │     sec/op      vs base                   │
GetMulti/size=1000B-24                                 1.344m ± 2%      1.296m ± 1%   -3.52% (p=0.000 n=80)
GetMulti/size=2KiB-24                                  2.610m ± 1%      2.424m ± 0%   -7.11% (n=80)
FindMissing/size=1000B-24                              485.2µ ± 2%      449.6µ ± 1%   -7.33% (p=0.000 n=80)
FindMissing/size=2KiB-24                               516.6µ ± 1%      444.1µ ± 2%  -14.04% (n=80)
Contains1/size=1000B-24                                9.004µ ± 1%      7.966µ ± 1%  -11.53% (n=80)
Contains1/size=2KiB-24                                 9.465µ ± 0%      7.959µ ± 1%  -15.92% (n=80)
Set/size=1000B-24                                      50.58µ ± 1%      50.18µ ± 2%        ~ (p=0.140 n=80)
Set/size=2KiB-24                                       193.6µ ± 6%      173.3µ ± 5%  -10.53% (p=0.000 n=78+80)
geomean                                                171.1µ           155.7µ        -8.97%

                          │ /home/lulu/benchmark-pebble-master.log │ /home/lulu/benchmark-pebble-get-proto.log │
                          │                  B/op                  │      B/op        vs base                  │
GetMulti/size=1000B-24                                541.7Ki ± 1%      521.1Ki ± 0%  -3.79% (n=80)
GetMulti/size=2KiB-24                                 902.1Ki ± 0%      874.7Ki ± 0%  -3.04% (n=80)
FindMissing/size=1000B-24                             298.3Ki ± 0%      298.1Ki ± 0%  -0.07% (p=0.002 n=80)
FindMissing/size=2KiB-24                              270.1Ki ± 0%      269.6Ki ± 0%  -0.19% (n=80)
Contains1/size=1000B-24                               3.578Ki ± 0%      3.369Ki ± 0%  -5.83% (n=80)
Contains1/size=2KiB-24                                3.358Ki ± 0%      3.086Ki ± 0%  -8.11% (n=80)
Set/size=1000B-24                                     8.382Ki ± 2%      8.059Ki ± 0%  -3.86% (p=0.000 n=80)
Set/size=2KiB-24                                      10.91Ki ± 1%      10.65Ki ± 1%  -2.41% (p=0.000 n=78+80)
geomean                                               50.64Ki           48.90Ki       -3.44%

                          │ /home/lulu/benchmark-pebble-master.log │ /home/lulu/benchmark-pebble-get-proto.log │
                          │               allocs/op                │    allocs/op     vs base                  │
GetMulti/size=1000B-24                                 7.934k ± 0%       7.733k ± 0%  -2.53% (n=80)
GetMulti/size=2KiB-24                                  11.56k ± 0%       11.36k ± 0%  -1.76% (n=80)
FindMissing/size=1000B-24                              6.016k ± 0%       6.014k ± 0%  -0.03% (n=80)
FindMissing/size=2KiB-24                               6.018k ± 0%       6.015k ± 0%  -0.05% (n=80)
Contains1/size=1000B-24                                 70.00 ± 0%        68.00 ± 0%  -2.86% (n=80)
Contains1/size=2KiB-24                                  70.00 ± 0%        68.00 ± 0%  -2.86% (n=80)
Set/size=1000B-24                                       128.0 ± 0%        126.0 ± 0%  -1.56% (n=80)
Set/size=2KiB-24                                        161.0 ± 1%        158.0 ± 1%  -1.86% (p=0.000 n=78+80)
geomean                                                 872.4             857.6       -1.70%


```